Revert the merge of 3.3.0~pre2 in upstream branch
git-svn-id: svn://svn.debian.org/svn/pkg-samba/branches/samba/upstream@2200 fc4039ab-9d04-0410-8cac-899223bdd6b0

+++ examples/libsmbclient/smbwrapper/wrapper.c
@@ -1109,7 +1109,7 @@ int utimes(const char *name, const struct timeval *tvp)
         return (* smbw_libc.utimes)((char *) name, (struct timeval *) tvp);
 }
 
-ssize_t readlink(const char *path, char *buf, size_t bufsize)
+int readlink(const char *path, char *buf, size_t bufsize)
 {
         check_init("readlink");
